refactor: split 890-line components.py into a package by concern
components.py held seven unrelated UI concerns; its own section comments ran
"Section 5" straight to "Section 13", which is accretion made visible. Split
into pulseengine/local/components/:

    sidebar.py       93   logo header, signal legend, top-mover rows
    scan_status.py   83   scan freshness: sidebar line + main-panel banner
    snapshot.py     129   signal card, why-box, metric cards
    news.py         125   article cards, clustered news section
    analysis.py     376   charts, signal breakdown, backtest, history
    overview.py     107   market heatmap, category overview table

Every module-level constant turned out to have exactly one consumer, so each
now lives beside its user and no shared _base module was needed. The
optional-import fallbacks sit in analysis.py, their only consumer. Code was
moved by line range rather than retyped, so the split cannot have silently
altered behaviour.

Both consumers are untouched and remain a documented interface:
  pulseengine/local/dashboard.py  ->  import ... as ui
  dashboard/components.py         ->  from ... import *

The public surface is 13 names, declared in __init__.__all__. The shim now
re-exports __all__ too, so it has its own declared export contract rather
than relying on whatever the star import happens to bind.

One silent-failure bug fixed in passing: _build_logo_html() walks up to the
repo root for assets/logo/, and moving it a directory deeper left parents[2]
pointing at pulseengine/. Nothing would have raised — the helper just returns
the emoji fallback and the dashboard quietly loses its logo. Corrected to
parents[3] and asserted.

Tests added rather than assumed:
- star-import through the shim, exercised with exec(), asserting the exported
  set equals an independently frozen list and that objects are identical
  (a list derived from __all__ would pass if a name were dropped from both)
- every ui.<name> the dashboard calls is in the public contract
- the logo resolves to the real PNG, not the fallback

Verified: ruff clean, mypy clean over 27 modules, 90 tests passing,
`python -m app.scan --dry-run` runs end to end, and from a wheel installed
into a throwaway venv the star import yields exactly the 13 names with
matching identity.
