+++ Docs/CHANGELOG.md
@@ -20,6 +20,8 @@ All notable changes to this project will be documented in this file.
 - `avg_signal_score` in `evaluate_signal_accuracy` (`pulseengine/core/backtest.py`) now computed as the mean of absolute scores, ensuring the reported average always reflects signal magnitude regardless of direction.
 
 ### Changed
+- **`pulseengine/local/components.py` (890 lines) split into a package.** It held seven unrelated UI concerns — its own section comments ran "Section 5" straight to "Section 13", visible accretion. Now `components/` with `sidebar.py`, `scan_status.py`, `snapshot.py`, `news.py`, `analysis.py` and `overview.py`, each 83–376 lines. Every module-level constant turned out to have exactly one consumer, so each lives with its user and no shared `_base` module was needed. Code was moved by line range, not retyped. Both consumers are unchanged: `import pulseengine.local.components as ui` and `from pulseengine.local.components import *`. The 13-name public surface is declared in `__init__.__all__` and frozen independently in `tests/test_legacy_interface.py`, which exercises real star-import semantics through the shim and asserts object identity.
+- **`_build_logo_html()` parent depth corrected during that split.** It walks up to the repo root for `assets/logo/`; moving it one directory deeper would have left `parents[2]` pointing at `pulseengine/`, and the failure is silent — the helper returns the emoji fallback and nothing raises. Now `parents[3]`, asserted in `tests/test_local_surface.py`.
 - **The dashboard stylesheet moved out of Python.** `pulseengine/local/styles.py` was 589 lines, 571 of them CSS inside a triple-quoted string. The CSS now lives in `pulseengine/local/dashboard.css` and `styles.py` is 36 lines that read it through `importlib.resources` (not a `__file__`-relative path, which would work in a checkout and silently lose all styling once installed). Editors syntax-highlight it, and CSS lines are no longer subject to Python line-length rules — which is what made 7 of the line-length findings below unfixable without wrecking a deliberately tabular block of signal colours.
 - **Optional-import fallback stubs in `pulseengine/local/components.py` now mirror the real signatures.** The three fallbacks used for `evaluate_signal_accuracy`, `get_signal_streak` and `get_historical_features` were declared `(*_a, **_kw)`, so they accepted calls the real functions reject — a bad call site would have worked only on machines where the optional import happened to fail. Surfaced by widening mypy past `core/`.
 - CI type-check scope widened from `mypy pulseengine/core/` to `mypy pulseengine/`; `local/` and `web/` were never checked, which is how the fallback mismatch above survived. Tests and the compat shims stay out of scope deliberately.
