+++ pyproject.toml
@@ -34,6 +34,7 @@ packages = [
     "pulseengine",
     "pulseengine.core",
     "pulseengine.local",
+    "pulseengine.local.components",
     "pulseengine.web",
     "src",
     "storage",
